Refactor/api integration viewmodel overhaul (#102)

+++ app/src/main/java/com/leic52dg17/chimp/domain/utils/formatHour.kt
@@ -4,8 +4,8 @@ import java.math.BigInteger
 import java.text.SimpleDateFormat
 import java.util.Date
 
-fun formatHours(timestamp: BigInteger): String {
-    val millis = timestamp.toLong() * 1000
+fun formatHours(timestamp: Long): String {
+    val millis = timestamp * 1000
     val date = Date(millis)
     val format = SimpleDateFormat("HH:mm")
     return format.format(date)
